Refactor status badge to common components.
