+++ explorer/app/components/Project/common/constants.ts
@@ -4,12 +4,3 @@
 export const enum CONTRIBUTOR_ROLE {
   DATA_CURATOR = "data curator",
 }
-
-/**
- * Possible set of project status values.
- */
-export enum STATUS {
-  NEW = "NEW",
-  NONE = "NONE",
-  UPDATED = "UPDATED",
-}
