+++ explorer/app/components/Project/common/entities.ts
@@ -1,10 +1,6 @@
 /**
  * Model of project collaborating organization to be used as props for the CollaboratingOrganizations component.
  */
-
-// App dependencies
-import { STATUS } from "./constants";
-
 export interface CollaboratingOrganization {
   citation: number;
   name: string;
@@ -57,11 +53,6 @@ export interface Publication {
   publicationTitle: string;
   publicationUrl: string;
 }
-
-/**
- * Model of project status to be used as props for the Hero component.
- */
-export type Status = keyof typeof STATUS;
 
 /**
  * Model of project supplementary link to be used as props for the SupplementaryLinks component.
