+++ explorer/app/components/Project/components/Hero/hero.stories.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 
 // App dependencies
-import { STATUS } from "../../common/constants";
+import { STATUS } from "../../../common/StatusBadge/statusBadge";
 import { Hero } from "./hero";
 
 export default {
